Time: 0 ms (100.00%) | Memory: 8.6 MB (60.84%) - LeetSync

+++ 4168-mirror-distance-of-an-integer/mirror-distance-of-an-integer.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int mirrorDistance(int n) {
-        long c;
-        long rev=0;
-        long temp = n;
+        int c;
+        int rev=0;
+        int temp = n;
         while(temp!=0){
             c = temp%10;
             rev = (rev*10) + c;
